UIC-ids568-milestone3/requirements.txt
# requirements.txt
apache-airflow==2.8.1
mlflow==2.8.1
scikit-learn==1.3.2
pandas==2.1.4
numpy==1.26.2
scipy==1.11.4
python-dotenv==1.0.0
argparse==1.4.0
requests==2.31.0
pyyaml==6.0.1